+++ GP2025/scenemain.cpp
@@ -47,7 +47,7 @@ bool SceneMain::Initialise(Renderer& renderer)
 
     m_pMineBackground->SetX(renderer.GetWidth() / 2);
     float scaledHeight = m_pMineBackground->GetHeight() * scale;
-    m_pMineBackground->SetY(scaledHeight / 2.0f);
+    m_pMineBackground->SetY((int)scaledHeight / 2);
     m_pMineBackground->SetScale(scale);
 
     m_pPlayer = new Player();
